Adding Javadocs and ALv2 headers in the files.
