--- shangridocs-services/src/main/java/org/shangridocs/services/wikipedia/Response/SectionInfo.java
@@ -1,24 +1,0 @@
-package org.shangridocs.services.wikipedia.Response;
-
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class SectionInfo {
-
-	private String line;
-	private String level;
-	public String getLine() {
-		return line;
-	}
-	public void setLine(String line) {
-		this.line = line;
-	}
-	public String getLevel() {
-		return level;
-	}
-	public void setLevel(String level) {
-		this.level = level;
-	}
-	
-	
-}
