--- shangridocs-services/src/main/java/org/shangridocs/services/wikipedia/Response/SubSectionResponse.java
@@ -1,19 +1,0 @@
-package org.shangridocs.services.wikipedia.Response;
-
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class SubSectionResponse {
-
-	ParsedResponse parse;
-
-	public ParsedResponse getParse() {
-		return parse;
-	}
-
-	public void setParse(ParsedResponse parse) {
-		this.parse = parse;
-	}
-	
-	
-}
