--- shangridocs-services/src/main/java/org/shangridocs/services/wikipedia/Response/WikipediaSearchResponse.java
@@ -1,21 +1,0 @@
-package org.shangridocs.services.wikipedia.Response;
-
-import java.util.Map;
-
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class WikipediaSearchResponse {
-	
-	Map<String,WikiPages> wikiPages;
-
-	public Map<String, WikiPages> getWikiPages() {
-		return wikiPages;
-	}
-
-	public void setWikiPages(Map<String, WikiPages> wikiPages) {
-		this.wikiPages = wikiPages;
-	}
-	
-	
-}
