+++ shangridocs-services/src/main/java/org/shangridocs/services/wikipedia/response/SectionInfo.java
@@ -1,0 +1,49 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.shangridocs.services.wikipedia.response;
+
+import org.codehaus.jackson.annotate.JsonIgnoreProperties;
+
+/** .... 
+ * 
+ * Sub class of SubSection Response :
+ * Sections details of the sections listed in Parsed Response
+ * 
+ * line : Subsection number of a sections in a page.
+ * level : Section Number in the page
+ * 
+ * */
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class SectionInfo {
+
+	private String line;
+	private String level;
+	public String getLine() {
+		return line;
+	}
+	public void setLine(String line) {
+		this.line = line;
+	}
+	public String getLevel() {
+		return level;
+	}
+	public void setLevel(String level) {
+		this.level = level;
+	}
+	
+	
+}
