+++ shangridocs-services/src/main/java/org/shangridocs/services/wikipedia/response/SubSectionResponse.java
@@ -1,0 +1,43 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.shangridocs.services.wikipedia.response;
+
+import org.codehaus.jackson.annotate.JsonIgnoreProperties;
+
+
+/** 
+ * 
+ *  This is the parent Object for subsection response
+ *  returned from WIKIPEDIA API. This object contains the bare minimum 
+ *  parsed data required by Shangridocs from the response returned by WIKI API.
+ *  
+ *  */
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class SubSectionResponse {
+
+	ParsedResponse parse;
+
+	public ParsedResponse getParse() {
+		return parse;
+	}
+
+	public void setParse(ParsedResponse parse) {
+		this.parse = parse;
+	}
+	
+	
+}
