+++ shangridocs-services/src/main/java/org/shangridocs/services/wikipedia/response/WikipediaSearchResponse.java
@@ -1,0 +1,42 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.shangridocs.services.wikipedia.response;
+
+import java.util.Map;
+
+import org.codehaus.jackson.annotate.JsonIgnoreProperties;
+
+/**
+ * Search Response Class, returned when a query is searched through the 
+ * Wiki database through a open search request to WIKI APIs.
+ *
+ */
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class WikipediaSearchResponse {
+	
+	Map<String,WikiPages> wikiPages;
+
+	public Map<String, WikiPages> getWikiPages() {
+		return wikiPages;
+	}
+
+	public void setWikiPages(Map<String, WikiPages> wikiPages) {
+		this.wikiPages = wikiPages;
+	}
+	
+	
+}
